requirements_update.txt
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.35.1
blinker==1.4
Bootstrap-Flask==1.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==7.1.1
Faker==4.0.2
Flask==1.1.2
Flask-Avatars==0.2.2
Flask-Dropzone==1.5.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
flask-whooshee==0.7.0
Flask-WTF==0.14.3
idna==3.10
isodate==0.7.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msrest==0.7.1
oauthlib==3.3.1
pathtools==0.1.2
Pillow==9.5.0
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.14.0
SQLAlchemy==1.3.15
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==0.10.2
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.2.1


